08-nodered/ctf/ctf_games/requirements.txt
paho-mqtt==1.6.1
python-dotenv==1.0.0
opcua==0.98.13
asyncua==1.0.0